Replace CategoricalConvPolicy
- Remove all occurrences of CategoricalConvPolicy
- Rename CategoricalConvPolicyWithModel to CategoricalConvPolicy
- Create and remove integration test
