+++ src/garage/tf/policies/__init__.py
@@ -2,8 +2,6 @@
 from garage.tf.policies.base import Policy
 from garage.tf.policies.base import StochasticPolicy
 from garage.tf.policies.categorical_conv_policy import CategoricalConvPolicy
-from garage.tf.policies.categorical_conv_policy_with_model import (
-    CategoricalConvPolicyWithModel)
 from garage.tf.policies.categorical_gru_policy import CategoricalGRUPolicy
 from garage.tf.policies.categorical_lstm_policy import CategoricalLSTMPolicy
 from garage.tf.policies.categorical_mlp_policy import CategoricalMLPPolicy
@@ -16,7 +14,6 @@ from garage.tf.policies.gaussian_mlp_policy import GaussianMLPPolicy
 
 __all__ = [
     'Policy', 'StochasticPolicy', 'CategoricalConvPolicy',
-    'CategoricalConvPolicyWithModel', 'CategoricalGRUPolicy',
     'CategoricalLSTMPolicy', 'CategoricalMLPPolicy', 'ContinuousMLPPolicy',
     'DiscreteQfDerivedPolicy', 'GaussianGRUPolicy', 'GaussianLSTMPolicy',
     'GaussianMLPPolicy'
